add tambak page + edit tambak
