+++ capstone-react-9/src/App.jsx
@@ -8,7 +8,6 @@ import AddArtikel from "./pages/Dashboard/AddArtikel";
 import DetailEditArtikel from "./pages/Dashboard/ArtikelDetail";
 import DataUser from "./pages/Dashboard/DataUsers";
 import Keluar from "./pages/Dashboard/Keluar";
-import Monitoring from "./pages/Dashboard/Monitoring";
 import Produk from "./pages/Dashboard/Produk";
 import ProdukEditdetail from "./pages/Dashboard/ProdukEditdetail";
 import Promo from "./pages/Dashboard/Promo";
@@ -31,7 +30,6 @@ function App() {
           <Route path="/artikel/detail/:id" element={<DetailEditArtikel />} />
           <Route path="/data-user" element={<DataUser />} />
           <Route path="/keluar" element={<Keluar />} />
-          <Route path="/monitoring" element={<Monitoring />} />
           <Route path="/produk" element={<Produk />} />
           <Route path="/produk/detail/:id" element={<ProdukEditdetail />} />
           <Route path="/promo" element={<Promo />} />
