--- capstone-react-9/src/components/AdminPage/MonitoringPage/index.jsx
@@ -1,14 +1,0 @@
-import LayoutDashboard from "../../LayoutDashboard/LayoutDashboard";
-import MonitoringMain from "./MonitoringMain";
-
-
-const MonitoringPage = () => {
-
-    return(
-        <LayoutDashboard>
-            <MonitoringMain/>
-        </LayoutDashboard>
-    )
-}
-
-export default MonitoringPage
