--- capstone-react-9/src/pages/Dashboard/Monitoring.jsx
@@ -1,5 +1,0 @@
-import MonitoringPage from "../../components/AdminPage/MonitoringPage";
-
-export default function Monitoring () {
-    return <MonitoringPage/>
-}
